+++ app/controllers/redmine_telegram_setup_controller.rb
@@ -1,65 +1,32 @@
 class RedmineTelegramSetupController < ApplicationController
+  include TelegramCommon::Tdlib::DependencyProviders::Authenticate
+
   unloadable
 
   def step_1
   end
 
   def step_2
-    result = telegram.execute('Login',
-      args: {
-          phone_number: params['phone_number']
-      }
-    )
-
-    params[:phone_code_hash] = JSON.parse(result)['phone_code_hash']
-
-    fail if params[:phone_code_hash].blank?
-  rescue => e
-    logger.fatal 'Failed to process API request'
-    logger.fatal e.to_s
-    logger.fatal result
-
-    msg = t('telegram_common.client.authorize.failed')
-
-    if e.class.to_s == 'TelegramCommon::Exceptions::Telegram'
-      msg = "#{msg}: #{e.to_s}"
+    begin
+      authenticate.(params)
+    rescue TelegramCommon::Tdlib::Authenticate::AuthenticationError => e
+      redirect_to plugin_settings_path('redmine_telegram_common'), alert: e.message
     end
-
-    return redirect_to plugin_settings_path('redmine_telegram_common'), alert: msg
   end
 
   def authorize
-    result = telegram.execute('Login',
-      args: {
-        phone_number:     params['phone_number'],
-        phone_code_hash:  params['phone_code_hash'],
-        phone_code:       params['phone_code']
-      }
-    )
-
-    fail if result != 'true'
-
-    save_phone_settings(phone_number: params['phone_number'])
-
-    return redirect_to plugin_settings_path('redmine_telegram_common'), notice: t('telegram_common.client.authorize.success')
-
-  rescue => e
-    logger.fatal 'Failed to process API request'
-    logger.fatal e.to_s
-    logger.fatal result
-
-    msg = t('telegram_common.client.authorize.failed')
-
-    if e.class.to_s == 'TelegramCommon::Exceptions::Telegram'
-      msg = "#{msg}: #{e.to_s}"
+    begin
+      authenticate.(params)
+      save_phone_settings(phone_number: params['phone_number'])
+      redirect_to plugin_settings_path('redmine_telegram_common'), notice: t('telegram_common.client.authorize.success')
+    rescue TelegramCommon::TdlibAuthenticate::AuthenticationError => e
+      redirect_to plugin_settings_path('redmine_telegram_common'), alert: e.message
     end
-
-    return redirect_to plugin_settings_path('redmine_telegram_common'), alert: msg
   end
 
   def reset
     save_phone_settings(phone_number: nil)
-    telegram.reset
+    FileUtils.rm_rf(Rails.root.join('tmp', 'redmine_telegram_common', 'tdlib'))
     redirect_to plugin_settings_path('redmine_telegram_common')
   end
 
@@ -67,9 +34,5 @@ class RedmineTelegramSetupController < ApplicationController
 
   def save_phone_settings(phone_number:)
     Setting.send('plugin_redmine_telegram_common=', Setting.plugin_redmine_telegram_common.merge({'phone_number' => phone_number.to_s}).to_h)
-  end
-
-  def telegram
-    @telegram ||= TelegramCommon::Telegram.new
   end
 end