+++ lib/telegram_common/tdlib.rb
@@ -1,0 +1,2 @@
+module TelegramCommon::Tdlib
+end