+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+
+# 0.2.0
+
+* Switch to TDlib
+
+# 0.1.4
+
+* Fix problem with telegram DC like AUTH_KEY_EMPTY
+
 # 0.1.3
 
 * Fixed bug with bot /connect command